+++ PluginsAndFeatures/azure-toolkit-for-rider/azure-intellij-plugin-bicep/src/main/kotlin/com/microsoft/azure/toolkit/intellij/bicep/lsp/BicepLanguageServerSupport.kt
@@ -1,60 +1,13 @@
 package com.microsoft.azure.toolkit.intellij.bicep.lsp
 
 import com.microsoft.azure.toolkit.intellij.bicep.BicepBundle.BICEP_EXTENSION
-import com.microsoft.azure.toolkit.intellij.bicep.settings.BicepConfigurable
 import com.intellij.execution.configurations.GeneralCommandLine
-import com.intellij.icons.AllIcons
 import com.intellij.openapi.editor.Editor
 import com.intellij.openapi.fileEditor.FileEditorManager
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.vfs.VirtualFile
-import com.intellij.platform.lsp.api.LspServer
 import com.intellij.platform.lsp.api.LspServerManager
-import com.intellij.platform.lsp.api.LspServerSupportProvider
-import com.intellij.platform.lsp.api.ProjectWideLspServerDescriptor
-import com.intellij.platform.lsp.api.customization.LspCustomization
-import com.intellij.platform.lsp.api.customization.LspFormattingCustomizer
-import com.intellij.platform.lsp.api.customization.LspFormattingSupport
-import com.intellij.platform.lsp.api.customization.LspInlayHintCustomizer
-import com.intellij.platform.lsp.api.customization.LspInlayHintDisabled
-import com.intellij.platform.lsp.api.lsWidget.LspServerWidgetItem
 import com.jetbrains.rider.NetCoreRuntime
-
-class BicepLspSupportProvider : LspServerSupportProvider {
-  override fun fileOpened(
-    project: Project,
-    file: VirtualFile,
-    serverStarter: LspServerSupportProvider.LspServerStarter
-  ) {
-    if (isLsAvailableFor(file, project)) {
-      serverStarter.ensureServerStarted(BicepLspDescriptor(project))
-    }
-  }
-
-  override fun createLspServerWidgetItem(lspServer: LspServer, currentFile: VirtualFile?): LspServerWidgetItem =
-    LspServerWidgetItem(lspServer, currentFile, AllIcons.Providers.Azure, BicepConfigurable::class.java)
-}
-
-class BicepLspDescriptor(project: Project) : ProjectWideLspServerDescriptor(project, "Bicep LSP") {
-  override fun isSupportedFile(file: VirtualFile): Boolean {
-    return file.extension == BICEP_EXTENSION
-  }
-
-  override fun createCommandLine(): GeneralCommandLine {
-    return prepareBicepServerLaunchCommandLine()
-  }
-
-  override val lspCustomization: LspCustomization = object : LspCustomization() {
-    override val inlayHintCustomizer: LspInlayHintCustomizer = LspInlayHintDisabled
-    override val formattingCustomizer: LspFormattingCustomizer = object : LspFormattingSupport() {
-      override fun shouldFormatThisFileExclusivelyByServer(file: VirtualFile,
-                                                           ideCanFormatThisFileItself: Boolean,
-                                                           serverExplicitlyWantsToFormatThisFile: Boolean): Boolean {
-        return file.extension == BICEP_EXTENSION
-      }
-    }
-  }
-}
 
 internal fun isLsAvailableFor(file: VirtualFile, project: Project): Boolean {
   if (file.extension != BICEP_EXTENSION) return false
