+++ PluginsAndFeatures/azure-toolkit-for-rider/azure-intellij-plugin-bicep/src/main/kotlin/com/microsoft/azure/toolkit/intellij/bicep/lsp/BicepLspSupportProvider.kt
@@ -1,0 +1,28 @@
+/*
+ * Copyright 2018-2025 JetBrains s.r.o. and contributors. Use of this source code is governed by the MIT license.
+ */
+
+package com.microsoft.azure.toolkit.intellij.bicep.lsp
+
+import com.intellij.icons.AllIcons
+import com.intellij.openapi.project.Project
+import com.intellij.openapi.vfs.VirtualFile
+import com.intellij.platform.lsp.api.LspServer
+import com.intellij.platform.lsp.api.LspServerSupportProvider
+import com.intellij.platform.lsp.api.lsWidget.LspServerWidgetItem
+import com.microsoft.azure.toolkit.intellij.bicep.settings.BicepConfigurable
+
+internal class BicepLspSupportProvider : LspServerSupportProvider {
+    override fun fileOpened(
+        project: Project,
+        file: VirtualFile,
+        serverStarter: LspServerSupportProvider.LspServerStarter
+    ) {
+        if (isLsAvailableFor(file, project)) {
+            serverStarter.ensureServerStarted(BicepLspDescriptor(project))
+        }
+    }
+
+    override fun createLspServerWidgetItem(lspServer: LspServer, currentFile: VirtualFile?): LspServerWidgetItem =
+        LspServerWidgetItem(lspServer, currentFile, AllIcons.Providers.Azure, BicepConfigurable::class.java)
+}
